+++ R/crew_launcher_slurm.R
@@ -157,6 +157,14 @@ crew_class_launcher_slurm <- R6::R6Class(
     #' @return Character vector of the lines of the job script.
     #' @param name Character of length 1, name of the job. For inspection
     #'   purposes, you can supply a mock job name.
+    #' @param attempt Positive integer, number of the current attempt.
+    #'   The attempt number increments each time a worker exits
+    #'   without completing all its tasks, and it resets
+    #'   back to 1 if a worker instance successfully completes
+    #'   all its tasks and then exits normally.
+    #'   By assigning vector arguments
+    #'   to the cluster-specific options of the controller,
+    #'   you can configure different sets of resources for different attempts.
     #' @examples
     #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
     #' launcher <- crew_launcher_slurm(
@@ -166,65 +174,66 @@ crew_class_launcher_slurm <- R6::R6Class(
     #' )
     #' launcher$script(name = "my_job_name")
     #' }
-    script = function(name) {
+    script = function(name, attempt) {
+      options <- crew_options_slice(private$.options_cluster, attempt)
       c(
         "#!/bin/sh",
         paste0("#SBATCH --job-name=", name),
         if_any(
-          is.null(private$.options_cluster$log_output),
+          is.null(options$log_output),
           character(0L),
-          paste0("#SBATCH --output=", private$.options_cluster$log_output)
+          paste0("#SBATCH --output=", options$log_output)
         ),
         if_any(
-          is.null(private$.options_cluster$log_error),
+          is.null(options$log_error),
           character(0L),
-          paste0("#SBATCH --error=", private$.options_cluster$log_error)
+          paste0("#SBATCH --error=", options$log_error)
         ),
         if_any(
-          is.null(private$.options_cluster$memory_gigabytes_required),
+          is.null(options$memory_gigabytes_required),
           character(0L),
           sprintf(
             "#SBATCH --mem=%sM",
             as.integer(
               round(
-                x = 1024 * private$.options_cluster$memory_gigabytes_required,
+                x = 1024 * options$memory_gigabytes_required,
                 digits = 0L
               )
             )
           )
         ),
         if_any(
-          is.null(private$.options_cluster$memory_gigabytes_per_cpu),
+          is.null(options$memory_gigabytes_per_cpu),
           character(0L),
           sprintf(
             "#SBATCH --mem-per-cpu=%sM",
             as.integer(
               round(
-                x = 1024 * private$.options_cluster$memory_gigabytes_per_cpu,
+                x = 1024 * options$memory_gigabytes_per_cpu,
                 digits = 0L
               )
             )
           )
         ),
         if_any(
-          is.null(private$.options_cluster$cpus_per_task),
+          is.null(options$cpus_per_task),
           character(0L),
           paste0(
             "#SBATCH --cpus-per-task=",
-            private$.options_cluster$cpus_per_task
+            options$cpus_per_task
           )
         ),
         if_any(
-          is.null(private$.options_cluster$time_minutes),
+          is.null(options$time_minutes),
           character(0L),
-          paste0("#SBATCH --time=", private$.options_cluster$time_minutes)
+          paste0("#SBATCH --time=", options$time_minutes)
         ),
         if_any(
-          is.null(private$.options_cluster$partition),
+          is.null(options$partition),
           character(0L),
-          paste0("#SBATCH --partition=", private$.options_cluster$partition)
+          paste0("#SBATCH --partition=", options$partition)
         ),
-        private$.options_cluster$script_lines
+        options$script_lines
       )
     }
   )