+++ R/utils_options.R
@@ -12,3 +12,11 @@ crew_options_validate.default <- function(options) {
     )
   )
 }
+
+crew_options_slice <- function(options, index) {
+  exclude <- names(formals(crew_options_cluster))
+  for (name in setdiff(names(options), exclude)) {
+    options[[name]] <- slice_bounded(options[[name]], index)
+  }
+  options
+}